With support for `isEnum` fixing 'expanding_generics_test.dart' was easy.
However, it breaks a rule that we have been following all the time:
It makes a distinction in the test between the behavior in transformed
code and the behavior in untransformed code: In untransformed code it
is possible to perform `reflectType` on an instance of a generic class
(e.g., on `D<int>`), but that cannot be done in transformed code now
and it is not likely to be supported unless we get some extra primitive
operations that will allow us to check things like `T <= S` where `T`
and `S` are values of type `Type`.

The situation is under control because we have `canReflectType`, and
the two modes will dutifully report that they can resp. cannot do
it, which means that the test will run on identical code for both
modes, but it will enter the body of one `if` statement in the
untransformed mode, and it will skip that body in transformed mode.

Is that unacceptable? Should we insist that untransformed code must
also reject to do `reflectType` here? If we do get some extra
primitives (or if we find a way to do it without them), we will break
the test anyway, so the careful programmer will generally have to
write code that checks `canReflectType` rather than just jumping to
`reflectType`.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1433363002 .