Adds support for type variables.
New test 'type_variables_test.dart' shows the new features at work.

Added more strict checking in pre-transform mode (adding method
`reflectableSupportsType` and executing more checks of this nature in
several locations), such that the two modes can perform identically,
and with an appropriate semantics.

Type variables are supported iff we have `declarationsCapability`,
that is, a type variable is considered to be a declared entity,
which matches the fact that it introduces a new name and may have
a "type" (it belongs to the set of subtypes of the given bound,
if any). Getting the (actual) type arguments from an instantiated
generic type also requires `declarationCapability`, for the same
reason. We don't support actually returning them, except when we
know that it is an empty list (i.e., the class has no type
arguments), but we can die correctly in the two cases (no capability
vs. _unsupported).

Adjusted `classMirrors` in reflectable data to be of type
`List<TypeMirror>` and renamed it to `typeMirror`, such that it can
hold type variable mirrors as well as class mirrors. Added code to
generate the type variable mirrors, find indexes, look them up,
etc.

Decided that `TypeAnnotationQuantifyCapability` is the right place to
ask for `upperBound` support. Without that, there is no capability
to ask for the upper bound of a type variable, with it we both get
the capability to ask and the coverage for the relevant classes.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1443703002 .