Checks for entry point existence and supports warning suppression.
We've had issue #17 forever, because we have no way to check whether a certain entry point actually exists (we do not know the root directory of a package when it is transformed).

This CL checks whether a specified entry point file exists, which eliminates all the warnings previously given for entry points in the local package (the one where `pub build..` or `pub serve..` is executed) because of the arguments given to pub (we would warn about missing 'test/*.dart' when running `pub build web`, because pub does not deliver 'test/*.dart' to the transformer in that case).

However, this approach used to fail for transformation of other packages (the ones that the local package depends on), because the root directory of such a package is unknown (and it is not the current working directory of the process that runs the transformer). However, I have made many attempts to reproduce this problem, and it never occurred. It is possible that `pub` has been modified.

On top of that, this CL provides support for suppression of warnings, using a setting in 'pubspec.yaml'.

Those two elements together should make it possible for users of this package to avoid the Info message about 'Missing entry point', and we may then close the issue.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1448333004 .