Fixes several incompleteness/bug problems, based on coverage.
In order to be able to recognize locations in code where invariants
have been violated (that is, locations which should never be reached),
this CL changes `_unsupported` and `throw new UnimplementedError` and
similar constructs, such that they all use the same, few variants:
`unreachableError` (indicating that this location _should_ be
anreachable, and it is a bug when it is reached) and
`unimplementedError` (indicating that it is OK to reach this location,
but the implementation of the behavior needed at that point is not
yet available).

Then the coverage support was used to obtain a list of unexecuted lines.
The unexecuted lines were categorized, and 21 unexecuted locations in
the code corresponding to missing test cases were detected (that is,
a supposedly real implementation was available there, but it had never
been tested). Various test cases were adjusted such that each of those
locations are reached during `make test`, and many bugs were fixed
along the way.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1465903002 .