Indexifies `reflectedType`, `dynamicReflectedType`, adds `bestEffort..`.
At first, `..eflectedType` was supported by a direct representation in
class mirrors, but type annotations caused them to be duplicated, so
it's better to keep them in a separate set and use indices. This CL
makes that happen.

It also adds `bestEffortReflectedType` and `hasBestEffortReflectedType`
which will deliver the `reflectedType` if present, otherwise try
`dynamicReflectedType`, and finally fail if that's also not available.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1476423003 .