Added `delegate` implementation and `Delegate..Capability`.
In order to support `delegate` in transformed code, we must translate
a given `Symbol` to the corresponding `String`, albeit only in cases
where that `Symbol` corresponds to a covered member. So this CL adds
the capability to create a mapping from symbols for the names of all
covered members to the corresponding strings, and uses this to
implement `delegate`. We may be able to use the same mapping to offer
better support for `noSuchMethod` but this has not been explored;
note that we may wish to rename `Delegate..Capability` if the ability
to translate symbols to strings becomes visible at the surface, e.g.,
if we decide that it is so useful in its own right that we will
offer something like `getName` / `getSymbol` on `Reflectable`, as a
less complete version of the same methods on `MirrorSystem` in
'dart:mirrors'.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1487633002 .