Implements `isSubtypeOf` and `isAssignableTo`.
Note that this CL changes the semantics of `superclass` to strictly
follow the documentation: It is now required to have a
`TypeRelationsCapability` in order to perform `superclass`, even in
the cases where this yields a mixin application. We used to allow that
because such a class is otherwise inaccessible (there is no piece of
syntax which denotes that class, so we cannot get it using
`reflectType`, and there are no direct instances of such a class, so
we cannot get it using `reflect`). Moreover, we also require a
`SuperclassQuantifyCapability` in order to reach such a mixin
application, because it is an anomaly to include it implicitly.

Finally, there are several locations where pre-transformed code will
attempt `invoke`, `getField`, or `setField` on a 'dart:mirrors' mirror.
These locations have now been protected by a try/catch, such that they
can throw a NoSuchCapabilityError, hence being consistent with the
behavior of post-transformed code.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1485193002 .