Implements the new no-such-method semantics in untransformed code.
This CL extensively rewrites the code for invocation in untransformed
code, such that an invocation which fails with a no-such-method
situation will be detected in Dart code (so there will be no native
invocation of `noSuchMethod` and there will be no `NoSuchMethodError`
exception), and then a `ReflectableNoSuchMethodError` will be thrown.

The point is that this is possible in transformed code, and in
transformed code we do not want to use `new Symbol(..)` and hence we
cannot invoke `noSuchMethod` correctly. With this approach the
treatment of no-such-method events is consistent, and programmers can
catch `ReflectableNoSuchMethodError` and handle the situation as they
wish --- all the data normally available in an `Invocation` is
available in a `StringInvocation` (but `memberName` is a `String`).

In order to do this, we need a large amount of extra checking, such
that we can predict whether or not there will be a no-such-method
situation in any given reflectable invocation, i.e., this is rather
costly. (So we may want to make it optional, if some users prefer the
incorrect semantics where `noSuchMethod` is just allowed to happen).

Support needed for this: New methods `_rawDeclarations`, `_rawType`,
and `_rawMetadata`, needed in order to bypass the capability checks
associated with `declarations` etc. (We must call these methods for
every invocation, not just when the user has specified a
`typeCapability` or `declarationsCapability`).

One extra detail: The `declarations` methods are now caching their
results --- it will be much more crucial to do this now, because we
call these methods for every invocation, not just when the user
explicitly requests the declarations of a class or a library.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1512033002 .