Enforce use of `ReflectableNoSuchMethodError`.
This CL enforces that every test case tests that it catches a
`ReflectableNoSuchMethodError` (rather than the less specific supertype
`NoSuchCapabilityError` that they used to test for) in all the cases
where it performs a reflectable invocation that must fail. This also
means that it is tested that we throw the right exception in all those
cases --- and this CL contains some corrections to achieve that,
especially in the method `delegate`.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1528583002 .