Changes code generation to avoid many elements.
This CL changes code generation such that code for parameter list shape
information is only generated when it is required (if we have mirrors
for methods and parameters then we will use them instead --- which is
slower the first time, but the result cached in each class mirror, and
it saves space to do it this way, given that we must have the method
and parameter mirrors for other reasons).

It also changes code generation such that method mirrors and parameter
mirrors are omitted entirely if there is no `declarationCapability`,
which means that configurations where only invocation is used can be
quite concise. (Previously, we could avoid having the class mirrors in
these cases, but we still had mirrors for instance members because
it was allowed to invoke `instanceMembers` without
`declarationsCapability`, and they implicitly used the method and
parameter mirrors).

That was a surprising and not-well-motivated corner case, so Sigurd
and I decided to change it. Hence, it is no more possible to use
`instanceMembers` and `staticMembers` without a
`declarationsCapability`. This is more consistent, and it enables the
space-saving changes of this CL to be applied in the cases where it
would be expected.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1528013003 .