Updates several version constraints, adjusts code to work there.
This CL switches from analyzer <=0.26.1+14 to ^0.27.0, in response to
the fixes applied in analyzer 0.27.1 that enables us to run with that
version (versions 0.26.1+15 .. 0.27.0 caused null related exceptions,
because constant values were not resolved).

It also switches code_transformers to ^0.3.0, which is linked to the
analyzer version (when 0.3.1 or higher is used, the analyzer can be
at or above 0.27.0). As a consequence, reflectable depends on package
'test', which creates some problems because it contains non-test files
with names matching '*_test.dart', which causes 'dart.test' to fail
(because those files are considered to be test cases, but they have
no `main` so it fails immediately). This CL handles that problem by
adding entries to the '.status' files.

Some methods were removed from `MixinApplication`: They have been
removed from the supertype in analyzer.

The file 'setup_analyzer.dart' was unused. With the version changes
it fails to pass dartanalyzer, so it was removed rather than fixed.

A bug concerned with obtaining metadata during capability checking
was fixed (adding support for `ConstructorElementImpl`). Several
null related bugs were fixed (especially: making the reaction to
null valued results from analyzer calls more robust).

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1537673002 .