This CL makes `_findEntryPoints` eliminate duplicates, and it adjusts 'check_literal_transform_test.dart' to test the response to an entry point duplication.
The file test_reflectable/pubspec.yaml contained a duplicate entry point, and this turned out to be the immediate cause of the infinite loop that we have seen in certain versions of the analyzer. The underlying cause in the analyzer has not been resolved, but there is no need to have duplicates of entry points, and they will now be eliminated.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1542073002 .