Aligning reflectable with sdk issues number 18541, 19635, and 25388.
Those three sdk issues are concerned with situations where some parts
of 'dart:mirrors' have not yet been implemented in `dart2js`. This CL
checks that the situations are already working in transformed code
with reflectable, and also with untransformed code for the vm/dartium.

Doing this, a few bugs were uncovered and fixed.

Issue 18541: 'Constructors of native classes are not included in
mirrors': Created new tests 'new_instance_native_test.dart' and
'new_instance_html_test.dart' to explore this scenario.

In this particular case we decided to special case `List`, because its
nameless constructor takes an argument which has a different default
value at runtime than that which is reported by the analyzer (we
cannot use the general approach that works for normal Dart code,
because this particular declaration has non-standard semantics).

Issue 19635 'Mirrors: newInstance on constructor with optional
parameters result in crash when running as JavaScript': Created
'new_instance_optional_arguments_test.dart' in order to check that
this will work in reflectable (suitably ported).

Issue 25388: 'dart2js: ClassMirror.invoke named arguments are not
implemented': Checked, this one is already covered by existing
tests.

Fixed bug: We used to omit all constructors for abstract classes
when generating the function literals used to invoke them, but
this prevents reflective invocation of factory constructors in
abstract classes.

Fixed bug: Corrected many occurrences of
`reflectableNoSuchMethodError` which should be one of the other
methods with a similar name, and introduced a new method
`reflectableNoSuchConstructorError` which was missing. This
fixes the problem that some error messages used to contain the
wrong word (e.g., said 'method' where it should have said
'constructor').

R=floitsch@google.com, sigurdm@google.com

Review URL: https://codereview.chromium.org/1573103003 .