This CL uses coverage support to eliminate bugs and enhance tests.
Running the coverage measurement produced 14 source code locations
which were associated with non-trivial, non-tested functionality.
Marked them with printouts, enhanced tests, and debugged failures
until they were covered, except for a couple which turn out to be
difficult (probably there is no way to reach those points, but
the code there is benign, should they be reached anyway). Comments
indicating that a location may be dead code have been added, such
that future searches for unexecuted lines can filter out these
locations as "already handled, as far as possible".

The most tricky points were associated with the combination of
generic classes and mixins, and in particular anonymous mixin
applications using type arguments from the enclosing class are
not well defined (an issue has been submitted on the sdk because
there are language issues and vm implementation issues in this).

This CL uncovers one case where an anonymous mixin application
using type arguments is considered as a generic class, which is
currently not the case (so this is ready for the fix, but it is
currently unreachable code). Such places have been marked with a
comment.

This CL fixes a bug associated with `reflectedType` and
`dynamicReflectedType`: the computation of values of
`reflectedReturnTypeIndex` and `dynamicReflectedReturnTypeIndex`
for return types previously ignored the value of
`reflectedTypeRequested`, which again masked an incorrect test
case.

This CL enhances tests to ensure that `_isTypeVariableFree`
is called recursively.

This CL marks code as possibly dead for another case dealing with
anonymous mixin applications: In the generation of expressions
evaluating to [Type] values used as `reflectedType` or
`dynamicReflectedType`. It seems unlikely that such a mixin
application can ever occur in this situation, but the code
dealing with it is benign, should it still happen.

This CL marks code which handles non-[InterfaceType]s coming from
`classes` or `reflectedTypes` as likely dead. Probably there will
never be any, but the treatment will work.

This CL enhances 'multi_field_test.dart' to use more shapes of
metadata, such that more code is covered.

This CL enhances 'generic_mixin_test.dart' to use a named mixin
application and a generic class, covering more code.

This CL adds `reflectedTypeCapability` to 'parameter_test.dart',
such that it will not stop working now that this capability
is required (as it should be). It changes some types in the
same test in order to cover more code due to more deeply nested
type arguments.

Added new test illustrating the
https://github.com/dart-lang/sdk/issues/25344.

Corrected .status to account for the failure of the new test
(which is "by design").

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1567783003 .