Changed the transformer to use `Transform`, not `AggregateTransform`.
This CL changes `AggregateTransform` to `Transform` in type
annotations, deletes a wrapper class used to create a `Transform`
from a given `AggregateTransform`, and simplifies the implementation
in several locations because simple transforms are now used rather
than aggregate ones.

The transformer has been using `AggregateTransform` nearly since the
beginning, because we transformed many libraries. Since we switched
a new model whereby only one library is modified (a newly generated
one) and one is renamed (and all others left untouched), it has been
unnecessary to use `AggregateTransform`s. It has even been argued
that it significantly reduces the performance of `pub serve` to use
an aggregate transform, because `pub serve` is unable to avoid
re-running the transformation for unchanged files, that is, it
transforms everything as soon as anything changes. So we wanted to
make this switch for a long time.

Moreover, it seemed likely that this change could eliminated the
problem which makes 'check_literal_transform_test.dart' flaky.
(In fact, it fails every time with newer sdk's, but it is marked as
flaky in the '.status' file because we do not know exactly which sdk
version is being used on bots). Unfortunately, the change did not
fix this problem.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1579923005 .