Restore the ability to do coverage measurements.
This CL fixes the issue (without issue number) that we have had for
quite a while, where `pub upgrade` could not be executed in
'coverage_reflectable' because it would not terminate (at least not
within 12 hours, and possibly never).

Apparently, we have to use a bunch of dependency_overrides in order to
be able to use the observatory (many other combinations do not work,
at least, they just cause `pub upgrade` to loop). This CL introduces
an adjusted copy of all those path dependencies that the observatory
uses, and that enables us to run `pub upgrade` as well as running the
coverage executable itself. As before, this will only work on one host,
but it also preserves the property that anyone who wants to run the
coverage executable can make the adjustments in a similar setup.

The script now filters out `unimplementedError` and `unreachableError`,
and shows the code in the given line; this should make it a lot easier
to ensure full coverage, because many lines which should not be
covered anyway are now removed from the output.

BUG=
R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1543003004 .