Added checks for null `ReflectorData`.
When a reflector does not match anything at all we do not create an
entry for it in `data`. This caused a null error to arise as reported
in issue 64. With the change this case is detected, and we get a proper
error message instead.

This CL fixes issue #64.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1602643002 .