A 'master' version of 1587113002, adding a bit of function types.
https://codereview.chromium.org/1587113002/ is a CL where a tiny amount
of support for functions was added to v0.4.0 (resolving issue #61 in
the context of code from outside, hence tied to that version).  This CL
ports that fix plus related fixes to the current master branch.

The main point is that code is _not_ generated for fresh `typedef`s for
anonymous function types. They arise, e.g., in `foo(int f(int _)`
where `f` has a function type that has no syntactic denotation and in
particular no name, but still allows 'dart:mirrors' to return a
`FunctionTypeMirror`). Even though we could start generating some
`typedef`s for these types and return that from `reflectedType` and
`dynamicReflectedType`, it seems useless to start implementing that:
A `Type` value is essentially only useful when you can test it for
equality, and that is not possible when no expression that the
programmer can write will evaluate to that `Type`.

So this CL makes `hasReflectedType` return false for these types in
untransformed code and ensures that we get a NO_CAPABILITY_INDEX
for them in transformed code.

Addresses https://github.com/dart-lang/reflectable/issues/61.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1605893002 .