Eliminated the conflict with newer sdk/tools.
The 'mock_tests' have been failing with all newer versions of the sdk
since dart-sdk-1.14.0-dev.5.0, but worked fine with .4.0. Turns out
that `Platform.packageRoot` has been changed at exactly that boundary
such that the newer version of the vm return a 'file://' and in general
produce that value in some steps from the actual command line argument
given in '--package-root=..', where as earlier versions would just
return the actual value from the command line.

Wrote a few lines to compensate for this change, and it works again.

Given that our red bot on Windows has been caused by the inability to
delete the file .flaky.log, and given that we will now eliminate the
only flaky status entry we have, this might make the tests work on
Windows again. (Presumably, it just won't work to have a flaky test
on Windows, but we just won't be affected by it any more).

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1602193002 .