Check for generation of private names.
As pointed out during the discussion about issue 64, the transformer
does not consistently prevent the generation of private names, and as
a result we may have a seemingly successful transformation followed by
error messages from dart2js or the vm. This CL adds checks for the
occurrence of private names in all the cases where we generate code
that puts a prefix and a name together (e.g., when we generate the
keys in the maps of `ReflectorData`, which was the concrete place
where the problem arose).

The CL also uses the now-working mock transformation framework to test
that a few malformed libraries will generate the expected detection and
diagnostic message, including a case with a private reflector class.

This CL addresses a side issue from
https://github.com/dart-lang/reflectable/issues/64.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1606193002 .