Adds a stand-alone version of the transformer.
This CL adds a new 'reflectable/bin/reflectable_transformer.dart' and
'test_reflectable/too/reflectable_transformer' (which is a script that
sets up the arguments to run the former). This is useful in order to
be able to debug the transformer itself (it is accessible for a
debugger which is not the case when it is run by pub), and also in
order to manually force transformation to take place for a subset of
the entry points in 'pubspec.yaml' without editing 'pubspec.yaml'.
Hence, it is obviously convenient for developers of reflectable.

It may also be usable for others, and it shows how to extract the
transformer from the `pub` context and use it on its own.

It does require a setup where `pub build --mode=debug test` or
something similar has been executed already, such that the 'packages'
directory has been created, `transform_import.dart` executed, etc.

With that, it creates identical output for the current test cases,
so it should be rather faithful (after a bunch of exceptions from that
rule have been debugged).

There are no tests for `reflectable_transformer{,.dart}` at this
point. We may wish to add a run that actually runs `pub build..`,
then transforms all tests in 'test_reflectable' (again) using
`reflectable_transformer`, and then runs all tests, but it is not
obvious how many changes to the testing procedure we would need in
order to do that.

Note that the method `hasCapability` has been renamed to
`_hasCapability`, in order to make the mirror/transformed interfaces
identical. Otherwise we may have programs using names exported from
'package:reflectable/src/reflectable_mirror_based.dart' (we do have
tests like that), and they would have incorrect generated code after
the switch to ..transformer_based..

A couple of changes were made in 'pubspec.yaml', because they should
have been performed some time ago, but only the testing procedure
with `reflectable_transformer` revealed the inconsistencies. For
instance, we should transform 'mixin_application_test.dart' even
though it does run successfully.

Finally, 'test/three_files_dir' was created and 'three_files_aux.dart'
moved in there in order to test code that handles the situation where
test libraries use imports across directory boundaries with relative
paths.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1613833004 .