Corrected the treatment of reflectable.pause.on.exit.
The control flow in `TransformerImplementation.apply` was relatively
natural for an `AggregateTransform` where `apply` would be called once
for each transformation (e.g., one invocation of `pub build..`). After
the switch to a (normal) `Transform` it was barely working, and surely
misleading. In particular, it did not (always) work to run coverage
measurements, because the occasion where it ended up pausing such that
the debugger (i.e., coverage measurement utility) would get a chance
to perform its job would not be at the end of the transformation.

With some added tests, this suddenly made the coverage measurements
fail.

This CL changes the control flow in `apply` to be natural for a plain
`Transform`, and it changes the code that checks when the overall
transformation has been completed. This also makes the coverage
measurement work again, hopefully much more robustly.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1622573002 .