Exclude function type mirrors from `_TypeAnnotationsFixedPoint`.
Reflectable has had red bots since the bots started running the tools
and sdk version 1.15.0-dev.1.0 (it is confirmed locally that the change
from version 1.14.0-dev.7.2 reproduces the success/failure locally on
the desktop and Mac). Apparently, the cause is that 'dart:mirrors' in
the new version returns some function type mirrors that are not equal
to themselves, which causes a set to grow indefinitely such that some
tests get a timeout.

With the current (lack of) support for function type mirrors, it is
benign (and saves a bit of space/time) to simply exclude all function
type mirrors during the computation of the type annotation fixed
point, so this is what is done in this CL.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1678093002 .