Adds documentation to mirrors in response to issue 72.
Issue 72 showed that error messages emitted from `isSubtypeOf` and
`isAssignableTo` were misleading. This CL changes them to match the
actual range of situations. Furthermore, documentation was added to
'reflectable/lib/mirrors.dart' as suggested by the submitter of
issue 72; for uniformity (and for practical usability) documentation
was transferred from 'dart:mirrors' to these two methods plus all
others where the documentation was totally absent, and it was
adjusted to match the actual behavior of reflectable.

This CL also changes `superinterfaces` to take `typeRelationsCapability`
into account, because that change was very closely related to the
changes performed in the other two methods (but this one used to have no
checks at all).

Small adjustments were made to the generated code in order to enforce
the use of `typeRelationsCapability` ("not supported" is `[-1]`, as
usual).

Addresses issue #72.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1695453002 .