Uses analyzer ^0.27.2 rather than a fixed 0.27.1
The transformer is modified to handle the situation where a constant
turns out to be unresolved (where `evaluationResult` returns null),
which used to prevent us from using any other version of the analyzer
than 0.27.1. The approach is lazy and fine-grained in that resolution
of constants in a library is requested just before the invocation of
`evaluationResult`, and only if this has not already been done. As an
extra bonus, the transformation of the 64 tests in 'test_reflectable'
is now completed about 12% faster than previously. About one in seven
libraries do not get their constants resolved, which is presumably
the main reason for the speedup.

Fixes https://github.com/dart-lang/reflectable/issues/54.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1761423002 .