Fixes an issue with metadata of enum values.
This CL adds a new test showing that there is a crash when the
transformer encounters an enum value and makes an attempt to extract
its metadata (in fact, enum values cannot have metadata, only enum
classes as a whole can have it). The CL detects and handles this case.

Fixes issue 80.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1841993003 .