Resolves another null problem with constants (metadata).
In the context of the polymer-dart issue #690, the reflectable
transformer repeatedly reported 'This reflector does not match
anything'. It turned out to be caused by `evaluationResult` returning
null (this time for metadata), and that was again caused by a cache
that got stale because the resolution data was being flushed from the
underlying `LibraryElement`. This CL eliminates that cache such that
we resolve constants every time they are needed, not just the first
time per library. This causes performance to worsen (`make test`
takes about 5% longer), but the extra work appears to be required
so that is hard to avoid.

Fixes issue #82, presumably helps resolving polymer-dart issue 690.

BUG=
R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1843063003 .