Cleans up handling of type relations.
Adds a new test which tests that usage of mirror methods requiring a
`typeRelationsCapability` when such a capability is not present will
cause a `NoSuchCapabilityError` to be thrown whose associated message
mentions 'ypeRelationsCapability' (we don't care about T/t).

Changes the generation of mixin indices such that this constraint is
enforced even in the cases where the relevant mixin class is covered.

Changes the handling of several methods in 'reflectable_mirror_based'
and 'reflectable_transformer_based' such that they perform this check.
Changes the wording of several other exception messages for
uniformity.

Addresses issue #77.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1850613003 .