Adds generation of type arguments to generated code for lists/maps.
This CL adds generation of type arguments passed to literal constructors
of lists and maps, when present in the original source code (e.g., with
complex metadata specifications, generating code for values returned by
`ClassMirror.metadata`). We used to generate code where these type
arguments are always omitted.

The discrepancy can be difficult to discover, because the test framework
special cases `Set`, `List`, and possibly other types, and then applies
an element-by-element comparison (so `[1, 2]` and `[1, 2]` are
"expect-equals" even though they are not `==`, and so are
`const <int>[]` and `[]`, and `new Set()` and `new Set()`).

Hence, the issue only arises in our test cases if we use an explicit
`==` in the `expect` invocation (thus avoiding "expect-equals" entirely)
or if we use nested structures where the list/map objects that differ
are wrapped in an object (`expect` only performs structural comparison
on certain types: that certainly includes `List` and `Map`, but I don't
know if there are others).

The new test 'literal_type_arguments_test.dart' uses an explicit `==`,
because this clearly shows that an `==` is being used, and also because
it avoids depending on the details of "expect-equals".

Fixes issue 83.

BUG=
R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1895823002 .