Added documentation for required capabilities + deprecated things.
This CL extends the library dartdoc for 'reflectable.dart' with
information about the reflectable package and its relation to
dart:mirrors, and it adds information to the dartdoc comments on almost
all methods about required (API) capabilities.

It also deprecates `NameCapability`, `nameCapability`,
`ClassifyCapability`, and `classifyCapability`, based on the experience
so for which says that it is cheap to have them and never necessary to
avoid having them. So they are just enabled at all times.

Many methods had stronger checks inserted in order to make the
capability requirements uniform; e.g., `apply` requires an
`InstanceInvokeCapability` on `call`, `defaultValue` requires a
`DeclarationsCapability`, and several others.

Similarly, several methods now have more precise diagnostic messages.
E.g., `isAssignableTo` will report that there is no
`TypeRelationsCapability` if that is the culprit, and only fall back
to claim that there is no support for the types involved when that is
actually true.

R=sigurdm@google.com

Review URL: https://codereview.chromium.org/1948513002 .